# Conflicts:
#	.env
#	.gitignore

+++ src/main/java/com/jit/rec/recipetoria/service/IngredientService.java
@@ -1,48 +1,74 @@
 package com.jit.rec.recipetoria.service;
 
+import com.jit.rec.recipetoria.dto.IngredientDTO;
+import com.jit.rec.recipetoria.entity.ApplicationUser;
 import com.jit.rec.recipetoria.entity.Ingredient;
-import com.jit.rec.recipetoria.entity.NewIngredientRequest;
+import com.jit.rec.recipetoria.entity.Recipe;
+import com.jit.rec.recipetoria.exception.ResourceNotFoundException;
 import com.jit.rec.recipetoria.repository.IngredientRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.MessageSource;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Locale;
 
 @Service
 @RequiredArgsConstructor
 public class IngredientService {
 
     private final IngredientRepository ingredientRepository;
+    private final MessageSource messageSource;
 
-    public List<Ingredient> getAllIngredients() {
-        return ingredientRepository.findAll();
+    public Ingredient createIngredient(IngredientDTO newRecipeIngredientDTO, ApplicationUser applicationUser) {
+        Ingredient newRecipeIngredient = IngredientDTO.convertToIngredient(newRecipeIngredientDTO);
+        newRecipeIngredient.setApplicationUser(applicationUser);
+
+        return ingredientRepository.save(newRecipeIngredient);
     }
 
-    public void createIngredient(NewIngredientRequest newIngredientRequest) {
-        Ingredient ingredient = new Ingredient();
+    public Ingredient createIngredient(IngredientDTO newRecipeIngredientDTO, Recipe recipe) {
+        Ingredient newRecipeIngredient = IngredientDTO.convertToIngredient(newRecipeIngredientDTO);
+        newRecipeIngredient.setRecipe(recipe);
 
-        ingredient.setName(newIngredientRequest.getName());
-        ingredient.setAmount(newIngredientRequest.getAmount());
-        ingredient.setMeasurementUnit(newIngredientRequest.getMeasurementUnit());
-
-        ingredientRepository.save(ingredient);
+        return ingredientRepository.save(newRecipeIngredient);
     }
 
-    public Ingredient getIngredientById(Long ingredientId) {
-        return ingredientRepository.findById(ingredientId).orElseThrow(() -> new IllegalStateException("NOT FOUND"));
+    public IngredientDTO getIngredientById(Long ingredientId) {
+        Ingredient ingredient = ingredientRepository.findById(ingredientId)
+                .orElseThrow(() -> new ResourceNotFoundException(messageSource.getMessage(
+                        "exception.ingredient.notFound", null, Locale.getDefault())));
+
+        return IngredientDTO.convertToDTO(ingredient);
     }
 
-    public void updateIngredientById(Long ingredientId, Ingredient updatedIngredientInfo) {
-        Ingredient ingredientToBeUpdated = getIngredientById(ingredientId);
+    public IngredientDTO updateIngredientById(Long ingredientId, IngredientDTO updatedIngredientInfo) {
+        Ingredient ingredientToBeUpdated = ingredientRepository.findById(ingredientId)
+                .orElseThrow(() -> new ResourceNotFoundException(messageSource.getMessage(
+                        "exception.ingredient.notFound", null, Locale.getDefault())));
 
-        ingredientToBeUpdated.setName(updatedIngredientInfo.getName());
-        ingredientToBeUpdated.setAmount(updatedIngredientInfo.getAmount());
-        ingredientToBeUpdated.setMeasurementUnit(updatedIngredientInfo.getMeasurementUnit());
+        ingredientToBeUpdated.setName(updatedIngredientInfo.name());
+        ingredientToBeUpdated.setAmount(updatedIngredientInfo.amount());
+        ingredientToBeUpdated.setMeasurementUnit(updatedIngredientInfo.measurementUnit());
 
         ingredientRepository.save(ingredientToBeUpdated);
+
+        return IngredientDTO.convertToDTO(ingredientToBeUpdated);
     }
 
     public void deleteIngredientById(Long ingredientId) {
+        if (!ingredientRepository.existsById(ingredientId)) {
+            throw new ResourceNotFoundException(messageSource.getMessage(
+                    "exception.ingredient.notFound", null, Locale.getDefault()));
+        }
         ingredientRepository.deleteById(ingredientId);
+    }
+
+    List<Ingredient> getAllIngredientsByApplicationUser() {
+        ApplicationUser applicationUser =
+                (ApplicationUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        return ingredientRepository.findAllByApplicationUser(applicationUser);
     }
 }
